Исправить analyzer

+++ stats/analyzer/src/main/java/ru/practicum/stats/analyzer/service/UserActionService.java
@@ -7,6 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewm.stats.avro.UserActionAvro;
 import ru.practicum.stats.analyzer.mapper.UserActionMapper;
+import ru.practicum.stats.analyzer.model.UserActionId;
 import ru.practicum.stats.analyzer.repository.UserActionRepository;
 
 @Service
@@ -19,7 +20,11 @@ public class UserActionService {
     public void save(UserActionAvro avro) {
         log.info("Сохранинеия дайствия {} пользователя {} для события {}", avro.getActionType(),
                 avro.getUserId(), avro.getEventId());
-        repository.deleteById(UserActionMapper.mapAvroToKey(avro));
-        repository.save(UserActionMapper.mapAvroToEntity(avro));
+        UserActionId id = UserActionMapper.mapAvroToKey(avro);
+        if (!repository.existsById(id) ||
+            repository.findById(id).get().getScore() < UserActionMapper.convertActionToWeight(avro.getActionType())) {
+            repository.deleteById(UserActionMapper.mapAvroToKey(avro));
+            repository.save(UserActionMapper.mapAvroToEntity(avro));
+        }
     }
 }
